Lots of stuff
Edited MasterGameScript variables so communications between scripts is easier.
Fixed the cave entrance and exit.
Added a fade in and out for area changes.
The game is now one big map and not scene changes, the player simply teleports to a new area upon entrance trigger. This was done to keep info between areas.
Sword angle of attack has been increased.